+++ player.c
@@ -154,7 +154,7 @@ void player_move_fleet(
         if (f_src->power == 0) {
             int index = get_index_in_list(player->fleets, f_src);
             assert(index >= 0);
-            player->fleets->remove_at(player->fleets, (unsigned int) index);
+            player->fleets->remove_at(player->fleets, index);
             galaxy->sectors[sx][sy]->fleet->destroy(galaxy->sectors[sx][sy]->fleet);
             galaxy->sectors[sx][sy]->fleet = NULL;
         }
@@ -201,7 +201,7 @@ void player_add_fleet(Player* player, Fleet* fleet)
     if (index < 0) {
         player->fleets->insert_end(player->fleets, fleet);
     } else {
-        player->fleets->insert_at(player->fleets, fleet, (unsigned int) index);
+        player->fleets->insert_at(player->fleets, fleet, index);
     }
 }
 
@@ -213,7 +213,7 @@ void player_remove_fleet(Player* player, Fleet* fleet)
 {
     int index = get_index_in_list(player->fleets, fleet);
     assert(index >= 0);
-    player->fleets->remove_at(player->fleets, (unsigned int) index);
+    player->fleets->remove_at(player->fleets, index);
 }
 
 
@@ -250,7 +250,7 @@ void player_add_planet(Player* player, Planet* planet)
     if (index < 0) {
         player->planets->insert_end(player->planets, planet);
     } else {
-        player->planets->insert_at(player->planets, planet, (unsigned int) index);
+        player->planets->insert_at(player->planets, planet, index);
     }
 }
 
@@ -263,7 +263,7 @@ void player_remove_planet(Player* player, Planet* planet, Galaxy* galaxy)
 {
     int index = get_index_in_list(player->planets, planet);
     assert(index >= 0);
-    player->planets->remove_at(player->planets, (unsigned int) index);
+    player->planets->remove_at(player->planets, index);
     if (get_player_index(player, galaxy) == 0) // notify human player that a planet has been lost
         notify_planet_lost(planet);
